fix CI bandit gate and address review findings
Fixes the red CI on this PR and the two real defects flagged in review.

CI failure (root cause):
- `validate_bandit_suppressions.py --require-active` rejected the inline
  comment `# nosec B110 - best-effort cleanup` in fetcher.py: the validator
  treats every token after `nosec` as a bandit rule ID, so `-`, `best-effort`
  and `cleanup` were invalid. Its non-zero exit aborted the job before the
  test step ran, which is why no pytest/mypy artifacts were produced and the
  downstream gates reported failure. Rationale moved to its own line.

Review findings:
- Go tester oversized-line recovery could spin without progress: a repeated
  LimitOverrunError from readuntil() was retried without discarding anything,
  so readline() met the same stuck buffer. Extracted `_drain_oversized_line`,
  which discards exactly `LimitOverrunError.consumed` bytes per iteration
  (bytes confirmed not to contain the separator) and bails out on EOF or a
  zero consumed count, guaranteeing forward progress.
- Producer sentinel delivery could strand a healthy consumer: the previous
  bounded wait gave up after 5s even on the normal completion path, though
  consumers only exit on the None marker and the pipeline awaits them all.
  Delivery now retries indefinitely while consumers are live, and only skips
  when `_cancel_all` has set `stop_event` (forced teardown, where consumers
  are cancelled directly and the queue would never drain).
- Aligned `verify_signature`'s docstring with its fail-closed behavior for a
  missing timestamp.
- Annotated new/modified test functions per the repo's typing rule, and
  widened `CircuitBreaker.recovery_timeout` to float (it is compared against
  time.monotonic() deltas).

Regression coverage: sentinel delivery under transient backpressure vs forced
teardown (the teardown test reproduces the original hang against pre-fix code),
and three drain-loop cases. flake8, black, mypy (436 files) and the bandit
suppression validator are clean; 1354 unit tests pass.

+++ src/configstream/circuit_breaker.py
@@ -12,7 +12,9 @@ class CircuitBreakerState(Enum):
 
 
 class CircuitBreaker:
-    def __init__(self, failure_threshold: int, recovery_timeout: int):
+    def __init__(self, failure_threshold: int, recovery_timeout: float):
+        # recovery_timeout is compared against time.monotonic() deltas, so it is
+        # inherently a float quantity of seconds; ints remain valid callers.
         self.failure_threshold = failure_threshold
         self.recovery_timeout = recovery_timeout
         self.failure_count = 0
